Create chatbot_main.py
Main program to run the Chatbot Capstone.  I made this original program made as final project for the Codecademy "Build Chatbots with Python" course.  

This retrieval-based chatbot is designed to talk to alien visitors about earth animals.  Ask it a question about an animal to see what it says!

Uses the NLTK, spaCy, re, and Numpy modules.